+++ CHANGES.md
@@ -8,6 +8,7 @@
 - Log out option works
 - Switch character option works
 - The "Remember me" checkbox works (rather, stops working when you leave it unchecked)
+- Status form works
 
 # 0.2.1
 - Fixed a nasty bug where the login menu didn't show up.